refactor: setQuantity increase, decrease 함수를 cartItem으로 올림
